perf(fp8): backward dgated 2-CTA + true-2CTA-MMA (E-gated) + default y1 fusion + int32 hardening
Backward dgated blockscaled GEMM (SM103/B300), permute-fused groupgemm:
- cluster_m=2 (2-CTA cooperative tiling) at E<=64: ~6-14% faster, bit-identical.
- tile_m=256 => true 2-CTA tcgen05 MMA (use_2cta_instrs) at E in [2,64]: a further
  -4.7% (E=64) / -5.2% (E=32) on the dgated kernel (ncu base-clock, H2048 I1024
  TK262144; tensor-pipe 38.8->40.7%), zero extra memory.
  num_experts==1 (weight-grad caller) stays tile_m=128; E>64 stays tile_m=128 cm=1.
- _DownProjection now threads the real num_experts into gemm_dgated.default_config.

Default-on y1 SwiGLU+FP8-quant up-proj epilogue fusion (alignment-gated on
TK%128==0 and I%128==0): eliminates the standalone y1 quant kernel (~132us/step)
and bf16 y1 HBM (~814 MiB/layer), precision-neutral; unaligned shapes fall back.

int32 hardening: int64 row indices in _pad_quantize_and_pack_kernel and
_dequant_blockscaled_fp8_kernel (rows*stride can exceed 2^31 on multi-GB tensors).

Safety (large-scale pretraining critical path), all on B300 SM103:
- tile_m=256 BIT-IDENTICAL to tile_m=128 (dz/colvec/y1s max_abs_diff=0): M-tiling
  preserves the K-reduction order.
- compute-sanitizer memcheck CLEAN (0 errors) at int32- AND uint32-overflow element
  offsets (max idx 2.21e9 and 6.44e9) under uniform + 80%-skew + single-dominant.
- New regression test tests/ops/test_dgated_true2cta_stress.py.

+++ sonicmoe/quack_utils/blockscaled_fp8_gemm.py
@@ -3936,7 +3936,8 @@ def _pad_quantize_and_pack_kernel(
     Avoids materializing a bf16 padded intermediate buffer.
     """
     row_base = tl.program_id(0) * BLOCK_ROWS
-    row_ids = row_base + tl.arange(0, BLOCK_ROWS)
+    # int64 row index to avoid row*stride overflow on multi-GB tensors (rows*K > 2^31).
+    row_ids = (row_base + tl.arange(0, BLOCK_ROWS)).to(tl.int64)
     row_mask_1d = row_ids < rows
 
     # Load source indices (invariant across groups)
